Change parseInt to parseFloat to allow for user input of decimals

+++ js/scripts.js
@@ -16,23 +16,22 @@ function quartsToGallons(number){
 $(document).ready(function(){
   $("form#cups-form").submit(function(event){
     event.preventDefault();
-    const cups = parseInt($("input#cups").val());
+    const cups = parseFloat($("input#cups").val());
     const result = cupsToPints(cups);
     $("#pints-result").text(result);
   });
 
   $("form#pints-form").submit(function(event){
     event.preventDefault();
-    const pints = parseInt($("input#pints").val());
+    const pints = parseFloat($("input#pints").val());
     const result = pintsToQuarts(pints);
     $("#quarts-result").text(result);
   });
 
   $("form#quarts-form").submit(function(event){
     event.preventDefault();
-    const quarts = parseInt($("input#quarts").val());
+    const quarts = parseFloat($("input#quarts").val());
     const result = quartsToGallons(quarts);
     $("#gallons-result").text(result);
   });
- 
 });
